+++ Classes/Extensions.UniversalConfig.cs
@@ -1,19 +1,14 @@
-﻿#pragma warning disable CS0162, CS1587, CS1591, CS1998, IDE0059, IDE0028
-
-/// <summary>
+﻿/// <summary>
 /// Author: Cornelius J. van Dyk blog.cjvandyk.com @cjvandyk
 /// This code is provided under GNU GPL 3.0 and is a copyrighted work of the
 /// author and contributors.  Please see:
 /// https://github.com/cjvandyk/Extensions/blob/main/LICENSE
 /// </summary>
 
-using Microsoft.Online.SharePoint.TenantAdministration;
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Text.Json;
 using static Extensions.Core;  //NuGet Extensions.cs
-using static System.Logit;     //Nuget Extensions.Logit
 
 namespace Extensions
 {
@@ -56,5 +51,3 @@ namespace Extensions
         }
     }
 }
-
-#pragma warning restore CS0162, CS1587, CS1591, CS1998, IDE0059, IDE0028