+++ Classes/System.Net.WebException.cs
@@ -1,6 +1,4 @@
-﻿#pragma warning disable CS1587, CS1998, IDE0059, IDE0028
-
-/// <summary>
+﻿/// <summary>
 /// Author: Cornelius J. van Dyk blog.cjvandyk.com @cjvandyk
 /// This code is provided under GNU GPL 3.0 and is a copyrighted work of the
 /// author and contributors.  Please see:
@@ -54,10 +52,12 @@ namespace Extensions
                         Convert.ToInt32(
                             ex.Response.Headers["Retry-After"]) + 60) * 1000);
                     //We add 60 seconds to the wait time for good measure.
+#pragma warning disable SYSLIB0014 // Type or member is obsolete
                     request = (System.Net.HttpWebRequest)System.Net.HttpWebRequest.Create(
                         ex.Response.ResponseUri.ToString()
                                                .TrimStart('{')
                                                .TrimEnd('}'));
+#pragma warning restore SYSLIB0014 // Type or member is obsolete
                     request.Method = method;
                     request.Accept = accept;
                     System.Net.WebResponse response;
@@ -78,4 +78,3 @@ namespace Extensions
         #endregion Retry()
     }
 }
-#pragma warning restore CS1587, CS1998, IDE0059, IDE0028